Generate SearchBuilder into the host application This gives plugins a place to add additional search methods without having an complex inheritance hierarchy.
