+++ lib/blacklight.rb
@@ -8,7 +8,8 @@ module Blacklight
   autoload :Configuration, 'blacklight/configuration'
   autoload :SearchFields, 'blacklight/search_fields'
   autoload :SearchBuilder, 'blacklight/search_builder'
-  
+  autoload :SearchBuilderBehavior, 'blacklight/search_builder_behavior'
+
   autoload :Document, 'blacklight/document'
   autoload :Solr, 'blacklight/solr'
