+++ lib/generators/blacklight/templates/search_builder.rb
@@ -1,0 +1,3 @@
+class SearchBuilder < Blacklight::SearchBuilder
+  include Blacklight::Solr::SearchBuilderBehavior
+end
